requirements.txt
unidecode
pillow